Fixed issues identified by Luke...   Account Information ·         Return Codes o   Success– 200 – The comment seems incorrect – no account was created. o   Bad Request – 400 – The comment seems incorrect.
Create Account
·         Request Body
o   Merchant ID – not shown as optional but should be
o   FirstName – shown as optional when it is not
o   LastName – shown as optional when it is not
·         Return Codes
o   Conflict – Code 400 is returned when trying to create an account with a duplicate email address
·         Response Body
o   FirstName – shown as optional when it is not
o   LastName – shown as optional when it is not
·         Errors
o   400 – The word ‘occured’ in the Detailed message is misspelled.
o   401 – The detailed message has two typos:
§  ‘… could not be parsed…”
§  ‘… is well formed and…”
o   403 – This error doesn’t occur when a duplicate email address is used.  A 404 error occurs instead.
o   404 – The detailed message for 404 errors doesn’t match the actual cause.
List Files
·         Response Body
o   File.ProjectID
§  Comment – ‘If not project…’ should be ‘If no project…’